fix: skip ssr for react-ace
