+++ gatsby-theme-oi-wiki/src/components/CodeEditor/Editor.tsx
@@ -1,0 +1,33 @@
+import React from 'react'
+import Ace from 'react-ace'
+import useDarkMode from '../../lib/useDarkMode'
+import { langModeMap } from '../../lib/play/codeLang'
+// add language to support here
+import 'ace-builds/src-noconflict/mode-c_cpp'
+// import 'ace-builds/src-noconflict/mode-java'
+import 'ace-builds/src-noconflict/mode-python'
+// import 'ace-builds/src-noconflict/mode-javascript'
+
+// light/dark theme
+import 'ace-builds/src-noconflict/theme-chrome'
+import 'ace-builds/src-noconflict/theme-tomorrow_night'
+
+import { CodeEditorProps } from '.'
+
+export default function Editor ({
+  lang,
+  ...aceProps
+}: Omit<CodeEditorProps, 'title'>): React.ReactElement {
+  const isDarkMode = useDarkMode()
+
+  return (
+    <Ace
+      theme={isDarkMode ? 'tomorrow_night' : 'chrome'}
+      width=""
+      height=""
+      fontSize={13}
+      mode={lang ? langModeMap[lang] : 'text'}
+      {...aceProps}
+    />
+  )
+}
